:recycle: separete handler.go to server and client events
