+++ usecases/service/ws/server.go
@@ -1,7 +1,12 @@
 package ws
 
 import (
+	"fmt"
+	"time"
+
 	"github.com/21hack02win/nascalay-backend/model"
+	"github.com/21hack02win/nascalay-backend/oapi"
+	"github.com/21hack02win/nascalay-backend/util/random"
 	"github.com/labstack/echo/v4"
 )
 
@@ -9,4 +14,458 @@ type RoomServer struct {
 	hub    *Hub
 	room   *model.Room
 	logger echo.Logger
+}
+
+// ROOM_NEW_MEMBER
+// 部屋に追加のメンバーが来たことを通知する (サーバー -> ルーム全員)
+func (s *RoomServer) sendRoomNewMemberEvent(room *model.Room) error {
+	if !s.room.GameStatusIs(model.GameStatusRoom) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventROOMNEWMEMBER,
+		Body: oapi.WsRoomNewMemberEventBody{
+			Capacity: room.Capacity.Int(),
+			HostId:   room.HostId.UUID(),
+			Members:  oapi.RefillUsers(room.Members),
+		},
+	})
+
+	return nil
+}
+
+// ROOM_UPDATE_OPTION
+// ゲームの設定を更新する (サーバー -> ルーム全員)
+func (s *RoomServer) sendRoomUpdateOptionEvent(body *oapi.WsRoomUpdateOptionEventBody) error {
+	if !s.room.GameStatusIs(model.GameStatusRoom) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventROOMUPDATEOPTION,
+		Body: body,
+	})
+
+	return nil
+}
+
+// GAME_START
+// ゲームの開始を通知する (サーバー -> ルーム全員)
+// ODAIフェーズを開始する
+func (s *RoomServer) sendGameStartEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusRoom) {
+		return errWrongPhase
+	}
+
+	for _, m := range s.room.Members {
+		c, ok := s.hub.userIdToClient[m.Id]
+		if !ok {
+			c.logger.Infof("client(userId:%s) not found", m.Id.UUID().String())
+			continue
+		}
+
+		s.sendMsgTo(c, &oapi.WsSendMessage{
+			Type: oapi.WsEventGAMESTART,
+			Body: &oapi.WsGameStartEventBody{
+				OdaiExample: random.OdaiExample(),
+				TimeLimit:   int(s.room.Game.TimeLimit),
+			},
+		})
+	}
+
+	// ODAIフェーズに移行
+	s.room.Game.Status = model.GameStatusOdai
+	// ODAIのカウントダウン開始
+	if !s.room.Game.Timer.Stop() {
+		<-s.room.Game.Timer.C
+	}
+	s.room.Game.Timer.Reset(time.Second * time.Duration(s.room.Game.TimeLimit))
+	s.room.Game.Timeout = model.Timeout(time.Now().Add(time.Second * time.Duration(s.room.Game.TimeLimit)))
+
+	go func() {
+		select {
+		case <-s.room.Game.Timer.C:
+			if err := s.sendOdaiFinishEvent(); err != nil {
+				s.logger.Error(s.sendEventErr(err, oapi.WsEventODAIFINISH).Error())
+			}
+		case <-s.room.Game.TimerStopChan:
+		}
+	}()
+
+	return nil
+}
+
+// ODAI_INPUT
+// お題入力が完了した人数を送信する (サーバー -> ルームの各員)
+func (s *RoomServer) sendOdaiInputEvent(readyNum int) error {
+	if !s.room.GameStatusIs(model.GameStatusOdai) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventODAIINPUT,
+		Body: &oapi.WsOdaiInputEventBody{
+			Ready: readyNum,
+		},
+	})
+
+	return nil
+}
+
+// ODAI_FINISH
+// 全員がお題の入力を完了したことor制限時間が来たことを通知する (サーバー -> ルーム全員)
+// クライアントはお題を送信する
+func (s *RoomServer) sendOdaiFinishEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusOdai) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventODAIFINISH,
+	})
+
+	return nil
+}
+
+// DRAW_START
+// キャンバス情報とお題を送信する (サーバー -> ルーム各員)
+func (s *RoomServer) sendDrawStartEvent() error {
+	// NOTE: 最後にお題を送信したユーザー(2回目以降は最後に絵を送信したユーザー)のクライアントから実行される
+	if !s.room.GameStatusIs(model.GameStatusDraw) {
+		return errWrongPhase
+	}
+
+	var (
+		game      = s.room.Game
+		drawCount = game.DrawCount
+	)
+
+	for _, o := range game.Odais {
+		if len(o.DrawerSeq) <= drawCount.Int() {
+			return errInvalidDrawCount
+		}
+
+		drawer := o.DrawerSeq[drawCount.Int()]
+		c, ok := s.hub.userIdToClient[drawer.UserId] // `drawCount`番目に描くユーザーのクライアント
+		if !ok {
+			s.logger.Infof("client(userId:%s) not found", drawer.UserId.UUID().String())
+			continue
+		}
+
+		drawnArea := make([]int, drawCount.Int())
+		for i := 0; i < drawCount.Int(); i++ {
+			drawnArea[i] = o.DrawerSeq[i].AreaId.Int()
+		}
+
+		s.sendMsgTo(c, &oapi.WsSendMessage{
+			Type: oapi.WsEventDRAWSTART,
+			Body: oapi.WsDrawStartEventBody{
+				AllDrawPhaseNum: s.room.AllDrawPhase(),
+				Canvas: oapi.Canvas{
+					AreaId:    drawer.AreaId.Int(),
+					BoardName: game.Canvas.BoardName,
+				},
+				DrawPhaseNum: game.DrawCount.Int(),
+				Img:          o.Img.AddPrefix(),
+				Odai:         o.Title.String(),
+				TimeLimit:    int(game.TimeLimit),
+				DrawnArea:    drawnArea,
+			},
+		})
+	}
+
+	return nil
+}
+
+// DRAW_INPUT
+// 絵を描き終えた人数を送信する (サーバー -> ルームの各員)
+func (s *RoomServer) sendDrawInputEvent(readyNum int) error {
+	if !s.room.GameStatusIs(model.GameStatusDraw) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventDRAWINPUT,
+		Body: &oapi.WsDrawInputEventBody{
+			Ready: readyNum,
+		},
+	})
+
+	return nil
+}
+
+// DRAW_FINISH
+// 全員が絵を完了したことor制限時間が来たことを通知する (サーバー -> ルーム全員)
+// クライアントは絵を送信する
+func (s *RoomServer) sendDrawFinishEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusDraw) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventDRAWFINISH,
+	})
+
+	return nil
+}
+
+// ANSWER_START
+// 絵が飛んできて，回答する (サーバー -> ルーム各員)
+func (s *RoomServer) sendAnswerStartEvent() error {
+	// NOTE: 最後にお題を送信した人のクライアントで行う
+	if !s.room.GameStatusIs(model.GameStatusAnswer) {
+		return errWrongPhase
+	}
+
+	s.hub.mux.Lock()
+	defer s.hub.mux.Unlock()
+
+	for _, v := range s.room.Game.Odais {
+		ac, ok := s.hub.userIdToClient[v.AnswererId]
+		if !ok {
+			s.logger.Infof("client(userId:%s) not found", v.AnswererId.UUID().String())
+			continue
+		}
+
+		s.sendMsgTo(ac, &oapi.WsSendMessage{
+			Type: oapi.WsEventANSWERSTART,
+			Body: oapi.WsAnswerStartEventBody{
+				Img:       v.Img.AddPrefix(),
+				TimeLimit: int(s.room.Game.TimeLimit),
+			},
+		})
+	}
+
+	// ANSWERのカウントダウン開始
+	s.room.Game.Timer.Reset(time.Second * time.Duration(s.room.Game.TimeLimit))
+	s.room.Game.Timeout = model.Timeout(time.Now().Add(time.Second * time.Duration(s.room.Game.TimeLimit)))
+
+	go func() {
+		select {
+		case <-s.room.Game.Timer.C:
+			if err := s.sendAnswerFinishEvent(); err != nil {
+				s.logger.Error(s.sendEventErr(err, oapi.WsEventANSWERFINISH).Error())
+			}
+		case <-s.room.Game.TimerStopChan:
+		}
+	}()
+
+	return nil
+}
+
+// ANSWER_INPUT
+// 回答の入力が完了した人数を送信する (サーバー -> ルームの各員)
+func (s *RoomServer) sendAnswerInputEvent(readyNum int) error {
+	if !s.room.GameStatusIs(model.GameStatusAnswer) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventANSWERINPUT,
+		Body: &oapi.WsAnswerInputEventBody{
+			Ready: readyNum,
+		},
+	})
+
+	return nil
+}
+
+// ANSWER_FINISH
+// 全員が回答の入力を完了したことor制限時間が来たことを通知する (サーバー -> ルーム全員)
+// クライアントは回答を送信する
+func (s *RoomServer) sendAnswerFinishEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusAnswer) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventANSWERFINISH,
+	})
+
+	return nil
+}
+
+// SHOW_START
+// 結果表示フェーズが始まったことを通知する (サーバー -> ルーム全員)
+func (s *RoomServer) sendShowStartEvent() error {
+	// NOTE: 最後に回答したユーザーのクライアントで行う
+	if !s.room.GameStatusIs(model.GameStatusShow) {
+		return errWrongPhase
+	}
+
+	s.resetBreakTimer()
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventSHOWSTART,
+	})
+
+	return nil
+}
+
+// SHOW_ODAI
+// 最初のお題を受信する (サーバー -> ルーム全員)
+func (s *RoomServer) sendShowOdaiEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusShow) {
+		return errWrongPhase
+	}
+	var (
+		game = s.room.Game
+		sc   = game.ShowCount.Int()
+	)
+	if len(game.Odais) < sc+1 {
+		return errNotFound
+	}
+
+	sender := oapi.User{}
+	for _, m := range s.room.Members {
+		if m.Id == game.Odais[sc].SenderId {
+			sender = oapi.RefillUser(&m)
+			break
+		}
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventSHOWODAI,
+		Body: &oapi.WsShowOdaiEventBody{
+			Sender: sender,
+			Next:   oapi.WsNextShowStatus("canvas"),
+			Odai:   game.Odais[sc].Title.String(),
+		},
+	})
+
+	return nil
+}
+
+// SHOW_CANVAS
+// 次のキャンバスを受信する (サーバー -> ルーム全員)
+func (s *RoomServer) sendShowCanvasEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusShow) {
+		return errWrongPhase
+	}
+
+	var (
+		game = s.room.Game
+		sc   = game.ShowCount.Int()
+	)
+
+	if len(game.Odais) < sc+1 {
+		return errNotFound
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventSHOWCANVAS,
+		Body: &oapi.WsShowCanvasEventBody{
+			Next: oapi.WsNextShowStatus("answer"),
+			Img:  game.Odais[sc].Img.AddPrefix(),
+		},
+	})
+
+	return nil
+}
+
+// SHOW_ANSWER
+// 最後の回答を受信する (サーバー -> ルーム全員)
+func (s *RoomServer) sendShowAnswerEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusShow) {
+		return errWrongPhase
+	}
+
+	var (
+		game = s.room.Game
+		sc   = game.ShowCount.Int()
+	)
+
+	if len(game.Odais) < sc+1 {
+		return errNotFound
+	}
+
+	var next oapi.WsNextShowStatus
+	if sc+1 < len(game.Odais) {
+		next = oapi.WsNextShowStatus("odai")
+	} else {
+		next = oapi.WsNextShowStatus("end")
+	}
+
+	answerer := oapi.User{}
+	for _, m := range s.room.Members {
+		if m.Id == game.Odais[sc].AnswererId {
+			answerer = oapi.RefillUser(&m)
+			break
+		}
+	}
+
+	var answer string
+	if a := game.Odais[sc].Answer; a != nil {
+		answer = a.String()
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventSHOWANSWER,
+		Body: &oapi.WsShowAnswerEventBody{
+			Answerer: answerer,
+			Next:     next,
+			Answer:   answer,
+		},
+	})
+
+	return nil
+}
+
+// NEXT_ROOM
+// ルームの表示に遷移する (サーバー -> ルーム全員)
+func (s *RoomServer) sendNextRoomEvent() error {
+	if !s.room.GameStatusIs(model.GameStatusRoom) {
+		return errWrongPhase
+	}
+
+	s.sendMsgToEachClientInRoom(&oapi.WsSendMessage{
+		Type: oapi.WsEventNEXTROOM,
+	})
+
+	return nil
+}
+
+// CHANGE_HOST
+// ホストが落ちた時に飛んできて，ホスト役を変更する (サーバー -> ルーム全員)
+func (s *RoomServer) sendChangeHostEvent() error {
+	room := s.room
+	found := false
+	s.hub.mux.Lock()
+	defer s.hub.mux.Unlock()
+
+	for _, v := range room.Members {
+		if _, ok := s.hub.userIdToClient[v.Id]; ok && v.Id != room.HostId {
+			found = true
+			room.HostId = v.Id
+			break
+		}
+	}
+
+	if !found {
+		return errNotFound
+	}
+
+	return nil
+}
+
+// BREAK_ROOM
+// 部屋が破壊されたときに通知する (サーバー -> ルーム全員)
+// 部屋が立ってからゲーム開始まで15分以上経過している場合，部屋を閉じる
+// このタイミングでサーバーは保持しているルームに関わる全データを削除
+func (s *RoomServer) sendBreakRoomEvent() error {
+	s.hub.mux.Lock()
+	defer s.hub.mux.Unlock()
+
+	for _, v := range s.room.Members {
+		if c, ok := s.hub.userIdToClient[v.Id]; ok {
+			s.hub.unregister(c)
+		}
+	}
+
+	if err := s.hub.repo.DeleteRoom(s.room.Id); err != nil {
+		return fmt.Errorf("failed to delete room: %w", err)
+	}
+
+	return nil
 }
